+++ src/Pim/Bundle/CatalogBundle/Entity/Attribute.php
@@ -35,94 +35,58 @@ class Attribute extends AbstractEntityAttribute implements
      */
     protected $options;
 
-    /**
-     * @var integer $sortOrder
-     */
+    /** @var integer $sortOrder */
     protected $sortOrder = 0;
 
-    /**
-     * @var AttributeGroup
-     */
+    /** @var AttributeGroup $group */
     protected $group;
 
-    /**
-     * @var boolean $useableAsGridColumn
-     */
+    /** @var boolean $useableAsGridColumn */
     protected $useableAsGridColumn;
 
-    /**
-     * @var boolean $useableAsGridFilter
-     */
+    /** @var boolean $useableAsGridFilter */
     protected $useableAsGridFilter;
 
-    /**
-     * @var $availableLocales ArrayCollection
-     */
+    /** @var ArrayCollection $availableLocales */
     protected $availableLocales;
 
-    /**
-     * @var $families ArrayCollection
-     */
+    /** @var ArrayCollection $families */
     protected $families;
 
-    /**
-     * @var integer $maxCharacters
-     */
+    /** @var integer $maxCharacters */
     protected $maxCharacters;
 
-    /**
-     * @var string $validationRule
-     */
+    /** @var string $validationRule */
     protected $validationRule;
 
-    /**
-     * @var string $validationRegexp
-     */
+    /** @var string $validationRegexp */
     protected $validationRegexp;
 
-    /**
-     * @var boolean $wysiwygEnabled
-     */
+    /** @var boolean $wysiwygEnabled */
     protected $wysiwygEnabled;
 
-    /**
-     * @var decimal $numberMin
-     */
+    /** @var decimal $numberMin */
     protected $numberMin;
 
-    /**
-     * @var decimal $numberMax
-     */
+    /** @var decimal $numberMax */
     protected $numberMax;
 
-    /**
-     * @var boolean $decimalsAllowed
-     */
+    /** @var boolean $decimalsAllowed */
     protected $decimalsAllowed;
 
-    /**
-     * @var boolean $negativeAllowed
-     */
+    /** @var boolean $negativeAllowed */
     protected $negativeAllowed;
 
-    /**
-     * @var datetime $dateMin
-     */
+    /** @var datetime $dateMin */
     protected $dateMin;
 
-    /**
-     * @var datetime $dateMax
-     */
+    /** @var datetime $dateMax */
     protected $dateMax;
 
-    /**
-     * @var string $metricFamily
-     */
+    /** @var string $metricFamily */
     protected $metricFamily;
 
-    /**
-     * @var string $defaultMetricUnit
-     */
+    /** @var string $defaultMetricUnit */
     protected $defaultMetricUnit;
 
     /**
@@ -131,14 +95,10 @@ class Attribute extends AbstractEntityAttribute implements
      */
     protected $maxFileSize;
 
-    /**
-     * @var array $allowedExtensions
-     */
+    /** @var array $allowedExtensions */
     protected $allowedExtensions;
 
-    /**
-     * @var integer
-     */
+    /** @var integer $minimumInputLength */
     protected $minimumInputLength = 0;
 
     /**
@@ -149,44 +109,28 @@ class Attribute extends AbstractEntityAttribute implements
      */
     protected $locale;
 
-    /**
-     * @var ArrayCollection $translations
-     */
+    /** @var ArrayCollection $translations */
     protected $translations;
 
-    /**
-     * @var boolean $required
-     */
+    /** @var boolean $required */
     protected $required;
 
-    /**
-     * @var boolean $unique
-     */
+    /** @var boolean $unique */
     protected $unique;
 
-    /**
-     * @var mixed $defaultValue
-     */
+    /** @var mixed $defaultValue */
     protected $defaultValue;
 
-    /**
-     * @var boolean $searchable
-     */
+    /** @var boolean $searchable */
     protected $searchable;
 
-    /**
-     * @var boolean $localizable
-     */
+    /** @var boolean $localizable */
     protected $localizable;
 
-    /**
-     * @var boolean $scopable
-     */
+    /** @var boolean $scopable */
     protected $scopable;
 
-    /**
-     * @var array
-     */
+    /** @var array $properties */
     protected $properties;
 
     /**