+++ src/Pim/Bundle/FilterBundle/DependencyInjection/PimFilterExtension.php
@@ -25,6 +25,7 @@ class PimFilterExtension extends Extension
     {
         $loader = new YamlFileLoader($container, new FileLocator(__DIR__ .'/../Resources/config'));
         $loader->load('form_types.yml');
+        $loader->load('utility.yml');
         $loader->load('filters.yml');
     }
 }