+++ src/Pim/Bundle/FilterBundle/Filter/Flexible/EntityFilter.php
@@ -5,7 +5,7 @@ namespace Pim\Bundle\FilterBundle\Filter\Flexible;
 use Doctrine\Common\Collections\ArrayCollection;
 use Oro\Bundle\FilterBundle\Filter\EntityFilter as OroEntityFilter;
 use Oro\Bundle\FilterBundle\Datasource\FilterDatasourceAdapterInterface;
-use Pim\Bundle\FlexibleEntityBundle\Manager\FlexibleManager;
+use Pim\Bundle\FilterBundle\Filter\ProductFilterUtility;
 
 /**
  * Flexible filter
@@ -41,8 +41,7 @@ class EntityFilter extends OroEntityFilter
 
         $this->util->applyFlexibleFilter(
             $ds,
-            $this->get(FilterUtility::FEN_KEY),
-            $this->get(FilterUtility::DATA_NAME_KEY),
+            $this->get(ProductFilterUtility::DATA_NAME_KEY),
             $this->extractIds($data['value']),
             $operator
         );
@@ -59,8 +58,7 @@ class EntityFilter extends OroEntityFilter
      */
     protected function getClassName()
     {
-        /** @var FlexibleManager $fm */
-        $flexibleManager = $this->util->getFlexibleManager($this->get(FilterUtility::FEN_KEY));
+        $flexibleManager = $this->util->getProductManeger();
         $valueName       = $flexibleManager->getFlexibleValueName();
         $valueMetadata   = $flexibleManager->getStorageManager()
             ->getMetadataFactory()